Combine guild Leadership and Settings pages. Metadata update now a modal.
