+++ packages/user/FractalCore/ui/src/pages/guilds/leadership/components/schedule-dialog.tsx
@@ -106,6 +106,7 @@ export const ScheduleDialog = ({
                 {trigger ?? (
                     <Button
                         size="sm"
+                        variant="outline"
                         onClick={() => {
                             setIsOpen(true);
                         }}
