+++ packages/user/FractalCore/ui/src/pages/guilds/leadership/index.tsx
@@ -4,6 +4,8 @@ import { GlowingCard } from "@shared/components/glowing-card";
 import { CardContent, CardHeader, CardTitle } from "@shared/shadcn/ui/card";
 
 import { RemoveRepItem } from "./components/remove-rep-item";
+import { SetEvaluationsScheduleItem } from "./components/set-evaluations-schedule-item.tsx";
+import { SetGuildMetadataItem } from "./components/set-guild-metadata-item";
 import { SetGuildRepItem } from "./components/set-guild-rep-item";
 
 export const Leadership = () => {
@@ -11,9 +13,11 @@ export const Leadership = () => {
         <PageContainer>
             <GlowingCard>
                 <CardHeader>
-                    <CardTitle>Leadership</CardTitle>
+                    <CardTitle>Guild Settings</CardTitle>
                 </CardHeader>
                 <CardContent className="space-y-4">
+                    <SetGuildMetadataItem />
+                    <SetEvaluationsScheduleItem />
                     <SetGuildRepItem />
                     <RemoveRepItem />
                 </CardContent>
